Fix for field unfocus widget when closing keyboard manually
