+++ lib/src/widgets/public.dart
@@ -1,7 +1,7 @@
-export 'automatic-unfocus.widget.dart';
 export 'base-list/public.dart';
 export 'nice-animated-fade-in.widget.dart';
 export 'nice-automatic-keep-alive-client.widget.dart';
+export 'nice-automatic-unfocus.widget.dart';
 export 'nice-error.widget.dart';
 export 'nice-loading-overlay.widget.dart';
 export 'nice-text-accent-punctuation.widget.dart';
